Fixing improperly scaled camera, gun, label and landing tags. The camera tags were causing some distance distortions in the game, right with the amount of scaling. See #3335.
Could mean interstellar travel on the cheap, but I didn't actually tried to fly and check if it changes the distance for real, or just the display. Althoug I might want to patent it, Elon Musk would scratch his head for a while. (Here you go detailed commit message, @laarmen :) )